Clean up Pytorch lightning GMM feature branch (#51)
* remove reference to different covariance types, and kmeans lightning

* remove kmeans metrics, refactor kmeans implementation to re-use code
